+++ src/components/SkipToMain/SkipToMain.module.scss
@@ -1,0 +1,28 @@
+@import '../../styles/theme';
+
+.link {
+  left: -999px;
+  position: absolute;
+  top: auto;
+  width: 1px;
+  height: 1px;
+  overflow: hidden;
+  z-index: -999;
+  &:active,
+  &:focus {
+    color: $white;
+    background-color: $dark;
+    left: auto;
+    top: auto;
+    width: 50%;
+    height: auto;
+    overflow: auto;
+    margin: 1rem 25%;
+    padding: 0.375rem 0;
+    border-radius: 1rem;
+    border: 0.375rem solid $dark;
+    text-align: center;
+    font-size: 1rem;
+    z-index: 999;
+  }
+}
