+++ src/components/SkipToMain/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import styles from './skip-to-main.module.scss'
+import styles from './SkipToMain.module.scss'
 
 export default () => (
   <a href="#main" className={styles.link}>
