+++ src/components/layout/index.js
@@ -3,8 +3,7 @@ import PropTypes from 'prop-types'
 import { Link } from 'gatsby'
 import { Layout, Menu } from 'antd'
 
-import SkipToMain from '../skip-to-main'
-// import Header from '../header'
+import SkipToMain from '../SkipToMain'
 import Breadcrumb from '../breadcrumb'
 
 import styles from './layout.module.scss'
